Ensure src package importable in tests
